Refactoring with Jotai
Still refactoring with Jotai. Most of the heavy lifting is done moving variables to global and shifting functions to local. Theme has been decided and barring constructive feedback will be the theme going forward. Currently working on Abilities.jsx. Need to focus on functionality. Theme not applying as cleanly as I had hoped. Overall may need a slight rework to avoid the squished feeling it has.